fix all the scroll issues without date
