+++ package.json
@@ -22,6 +22,7 @@
     "react": "^18.2.0",
     "react-datepicker": "^4.16.0",
     "react-dom": "^18.2.0",
+    "react-infinite-scroll-component": "^6.1.0",
     "recharts": "^2.8.0",
     "styled-components": "^5.3.11"
   },
