+++ src/App.tsx
@@ -7,6 +7,7 @@ import { selectOptions } from "./mockData/SelectOptionsDropDown";
 //import { articles } from "./mockData/Article";
 import { CssBaseline } from "@mui/material";
 import AppContext from "./context/AppContext";
+import InfiniteScroll from "react-infinite-scroll-component";
 
 import AppHeader from "./components/AppHeader/AppHeader";
 import {
@@ -36,73 +37,46 @@ import { SourcesArray } from "./mockData/DashboardMockData";
 import { Api, buildApiQuery } from "./services/Api";
 import { DEFAULT_COUNTRY, PAGE_SIZE } from "./utils/constants/Constants";
 import { Article } from "./models/ArticleInterface";
+import { useFetchArticles } from "./hooks/useFetchArticles";
+import { useSearch } from "./hooks/useSearch";
 
 function App() {
   const [dateRange, setDateRange] = useState<[Date | null, Date | null]>([
     null,
     null,
   ]);
+  const [firstRender, setFirstRender] = useState(true);
+  const [hasMore, setHasMore] = useState(true);
   const [uniqueSources, setUniqueSources] = useState<SelectOption[]>([]);
   const [articlesType, setArticlesType] =
     useState<ArticlesType>("Top Headlines");
-  const [searchInput, setSearchInput] = useState<string>("");
   const [selectedSortBy, setSelectedSortBy] = useState<string>("");
   const [selectedSource, setSelectedSource] = useState<string>("");
   const [selectedLanguage, setSelectedLanguage] = useState<string>("");
   const [selectedCountry, setSelectedCountry] = useState<string>("");
   const [selectedCategory, setSelectedCategory] = useState<string>("");
-  const [debouncedSearchInput, setDebouncedSearchInput] = useState(searchInput);
-  const [recentSearches, setRecentSearches] = useState<string[]>([]);
   const [articles, setArticles] = useState<Article[]>([]);
-  //const articles = ArticelsMock.articles;
   const [currentPage, setCurrentPage] = useState(1);
   const articlesBySource = countArticlesBySource(articles);
-
+  // const [debouncedSearchInput, setDebouncedSearchInput] = useState(searchInput);
+  // const [recentSearches, setRecentSearches] = useState<string[]>([]);
+  // const [searchInput, setSearchInput] = useState<string>("");
+  //const articles = ArticelsMock.articles;
+  const [shouldFetch, setShouldFetch] = useState<boolean>(true);
   useEffect(() => {
     const sources: SelectOption[] = createSourcesOptions(SourcesMock);
     setUniqueSources(sources);
   }, []);
 
-  useEffect(() => {
-    const timerId = setTimeout(() => {
-      setDebouncedSearchInput(searchInput);
-    }, 500);
+  const {
+    searchInput,
+    setSearchInput,
+    setDebouncedSearchInput,
+    debouncedSearchInput,
+    recentSearches,
+    setRecentSearches, // Include setRecentSearches
+  } = useSearch();
 
-    return () => {
-      clearTimeout(timerId);
-    };
-  }, [searchInput]);
-  useEffect(() => {
-    const fetchDefaultData = async () => {
-      try {
-        const response = await Api.get("top-headlines", {
-          params: {
-            country: DEFAULT_COUNTRY,
-            pageSize: PAGE_SIZE,
-          },
-        });
-        setArticles(response.data.articles);
-
-        // Reset state to initial values
-        setArticlesType("Top Headlines");
-        setSelectedCountry("");
-        //setSelectedCategory("");
-      } catch (error) {
-        console.error("Error fetching default data:", error);
-      }
-    };
-
-    fetchDefaultData();
-  }, []);
-
-  useEffect(() => {
-    if (debouncedSearchInput) {
-      setRecentSearches((prevSearches) => [
-        ...prevSearches,
-        debouncedSearchInput,
-      ]);
-    }
-  }, [debouncedSearchInput]);
   useEffect(() => {
     resetArticlesAndPage();
   }, [
@@ -112,54 +86,11 @@ function App() {
     selectedLanguage,
     selectedCountry,
     selectedCategory,
-    dateRange,
+    dateRange, ///// maybe change to should fetch all this dependncies
   ]);
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const options: any = {};
 
-        if (articlesType === "Top Headlines") {
-          if (selectedCountry) options.country = selectedCountry;
-          if (selectedCategory) options.category = selectedCategory;
-          if (selectedSource) options.sources = [selectedSource];
-          if (searchInput) options.q = searchInput;
-        } else if (articlesType === "Everything") {
-          const [startDate, endDate] = dateRange;
-
-          if (startDate) options.from = startDate.toISOString();
-          if (endDate) {
-            const adjustedEndDate = new Date(endDate);
-            adjustedEndDate.setDate(endDate.getDate() + 1);
-            options.to = adjustedEndDate.toISOString();
-          }
-          if (selectedLanguage) options.language = selectedLanguage;
-          if (selectedSortBy) options.sortBy = selectedSortBy;
-          if (selectedSource) options.sources = [selectedSource];
-          if (searchInput) options.q = searchInput;
-        }
-
-        const params = buildApiQuery(articlesType, options, currentPage);
-        params.pageSize = PAGE_SIZE;
-
-        const response = await Api.get(
-          articlesType === "Top Headlines" ? "top-headlines" : "everything",
-          { params }
-        );
-        console.log(params);
-
-        console.log(response.data);
-        setArticles((prevArticles) => [
-          ...prevArticles,
-          ...response.data.articles,
-        ]);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      }
-    };
-
-    fetchData();
-  }, [
+  useFetchArticles(
+    firstRender,
     debouncedSearchInput,
     selectedSortBy,
     selectedSource,
@@ -168,26 +99,44 @@ function App() {
     selectedCategory,
     dateRange,
     currentPage,
-  ]);
-  useEffect(() => {
-    const handleScroll = () => {
-      if (window.innerHeight + window.scrollY >= document.body.offsetHeight) {
-        setCurrentPage((prevPage) => prevPage + 1);
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
-
+    articlesType,
+    searchInput,
+    setArticles,
+    setFirstRender,
+    setHasMore,
+    shouldFetch
+  );
   //// make 1 sec of 0 articels
   const resetArticlesAndPage = () => {
-    setArticles([]);
-    setCurrentPage(1);
+    //if articles type change the prev articels still render.
+    if (
+      debouncedSearchInput !== "" ||
+      selectedSortBy !== "" ||
+      selectedSource !== "" ||
+      selectedLanguage !== "" ||
+      selectedCountry !== "" ||
+      selectedCategory !== "" ||
+      (dateRange[0] !== null && dateRange[1] !== null)
+    ) {
+      setArticles([]);
+      setCurrentPage(1);
+    }
   };
+
+  useEffect(() => {
+    setShouldFetch(false); // Prevent fetching
+    setSearchInput("");
+    setSelectedSortBy("");
+    setSelectedSource("");
+    setSelectedLanguage("");
+    setSelectedCountry("");
+    setSelectedCategory("");
+  }, [articlesType]);
+  useEffect(() => {
+    if (!shouldFetch) {
+      setShouldFetch(true);
+    }
+  }, [shouldFetch]);
 
   return (
     <>
@@ -227,7 +176,14 @@ function App() {
             <p> {articles.length} seraches</p>
             {/* )} */}
             <MainContent>
-              <ArticelsList atricleList={articles} />
+              <InfiniteScroll
+                dataLength={articles.length}
+                next={() => setCurrentPage((prevPage) => prevPage + 1)}
+                hasMore={hasMore} // You can set a condition to stop loading more articles
+                loader={<h4>Loading...</h4>}
+              >
+                <ArticelsList atricleList={articles} />
+              </InfiniteScroll>
               <Dashboard
                 data={{
                   totalArticles: articles.length,
